Implement sidebar & second method for calculating with reference object
Implement sidebar(navigation drawer) for switching between different methods. Implement a component for the second method, which takes a picture of a person with a QR code (of known size, containing size dimensions as data) on their forehead.

The image can only be taken once both the face and the QR code are in view. The image taken is then sent (POST) to the backend for processing and the app is awaiting a response.