Fix QR scanning issue where picture got taken without a QR code in view
Now performs a check on whether the QR code is still in view with every countdown tick.

Since the barcode scanning functionality only returns a value when a barcode is detected, we maek it set a value to true with every countdown tick to ensure the QR code is still in view.